fix: Add hokuyo counterbalance in the back
I'm trying to fix the car toppling forward on stop problem. I don't
think I'll be able to get rid of the problem completely. But I'm just
going to add things till it kinda sorta does not topple.

This may require that the car never speed up very much. I guess that is
okay for now.